+++ plugins/org.preesm.codegen.xtend/stdfiles/c/communication.h
@@ -50,6 +50,39 @@ communications.
 #define _PREESM_COMMUNICATION_H
 
 
+#ifdef __APPLE__
+#include <dispatch/dispatch.h>
+#else
+#include <semaphore.h>
+#endif
+
+struct rk_sema {
+#ifdef __APPLE__
+    dispatch_semaphore_t    sem;
+#else
+    sem_t                   sem;
+#endif
+};
+
+#ifdef _WIN32
+void rk_sema_init(struct rk_sema *s, int value);
+#else
+void rk_sema_init(struct rk_sema *s, int value);
+#endif
+
+#ifdef _WIN32
+void rk_sema_wait(struct rk_sema *s);
+#else
+void rk_sema_wait(struct rk_sema *s);
+#endif
+
+#ifdef _WIN32
+void rk_sema_post(struct rk_sema *s);
+#else
+void rk_sema_post(struct rk_sema *s);
+#endif
+
+
 /**
 * Maximum number of core supported by the communication library.
 * This number is used to allocate the table of semaphores used for intercore